First functional get-workshops.py script to pull get workships in ICLR.cc format.
